Reapply "M320-D2 TEV changes"
This reverts commit f095cf126156d0bc54cc6165c64ba13c9c572963.

+++ M320-D2/src/Test.java
@@ -26,6 +26,6 @@ public class Test {
     }
 
     public float calculateGrade() {
-        return ((points / maxPoints * 5) + 1);
+        return (((float) points / maxPoints * 5) + 1);
     }
 }
